Convert from solc 0.7.6 to 0.8.11
This gets us 90% of the way there. However there are three issues:
1) A couple of OZ files were copied in directly in order to get around multiple declration issues (hexString defined locally and in OZ)
2) INonFungiblePositionManager extending IERC721Enumerable was commented out in order to fix broken dependencies
3) "CompilerError: Stack too deep when compiling inline assembly: Variable pos is 19 slot(s) too deep inside the stack." but I can't figure out where "pos" is, and even if I did, it sounds like a nightmare to fix. It might be possible to fix by converting to abicoder v1 to get rid of some extra stack allocations, but we need to find "pos" first.

+++ contracts/libraries/PoolAddress.sol
@@ -33,13 +33,15 @@ library PoolAddress {
     function computeAddress(address factory, PoolKey memory key) internal pure returns (address pool) {
         require(key.token0 < key.token1);
         pool = address(
-            uint256(
-                keccak256(
-                    abi.encodePacked(
-                        hex'ff',
-                        factory,
-                        keccak256(abi.encode(key.token0, key.token1, key.fee)),
-                        POOL_INIT_CODE_HASH
+            uint160(
+                uint256(
+                    keccak256(
+                        abi.encodePacked(
+                            hex'ff',
+                            factory,
+                            keccak256(abi.encode(key.token0, key.token1, key.fee)),
+                            POOL_INIT_CODE_HASH
+                        )
                     )
                 )
             )
